qemu: implement device backends functionality
This allows the user to specify which backend should be used for
the disk and network devices. Currently, only the "network" and "disk"
devices are able to be specified. If they are not specified, QEMU will
use the default backends ("e1000" and "none", respectively).

An example of the usage of this could be setting the disk backend to be
"virtio", and the network backend to be "virtio-net-pci":

    backends:
      qemu:
        systems:
          - ubuntu-16.04:
              device-backends:
                disk: virtio
                network: virtio-net-pci

The configuration values are passed verbatim to the QEMU instance,
without checking if the values are supported by QEMU, allowing the
flexibility for the user to chose arbitrary backends.

In the case of the disk backend, the argument will be passed as the
value for the `if=` parameter of the `-drive` argument. For the network
device backend, the value will be passed as the value of the `driver=`
parameter of the `-device` argument.

Signed-off-by: Isaac True <isaac.true@canonical.com>

add test

Signed-off-by: Isaac True <isaac.true@canonical.com>